+++ Changelog.md
@@ -1,11 +1,23 @@
 # Release History
 
+
+
+### 2.0.0 ( 2023-09-29 )
+- [x] Refactoring: React-router is no longer a dependency;
+- [x] Complete rewrite of the application;
+
+
+
 ### 1.1.1 ( 2022-09-18 )
 - [x] Fix: Use of method 'repeat' before a navigation event is cause for error;
+
+
 
 ### 1.1.0 ( 2022-07-22)
 - [x] New methods: getCurrent, repeat;
 - [ ] Bug: Use of method 'repeat' before a navigation event is cause for error;
+
+
 
 ### 1.0.1 ( 2022-10-08 )
 - [x] Published for first time